backend/requirements.txt
django==4.2.7
django-tenants==3.6.1
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.10
django-cors-headers==4.6.0
gunicorn==23.0.0
